# Cleanup spec_helper.rb to speed up runs