--- a/moduleroot/spec/spec_helper.rb
+++ b/moduleroot/spec/spec_helper.rb
@@ -11,29 +11,8 @@ require 'augeas_spec'
 <% end -%>
 
 RSpec.configure do |c|
-  c.include PuppetlabsSpec::Files
-
-  c.before :each do
-    # Ensure that we don't accidentally cache facts and environment
-    # between test cases.
-    Facter::Util::Loader.any_instance.stubs(:load_all)
-    Facter.clear
-    Facter.clear_messages
-
-    # Store any environment variables away to be restored later
-    @old_env = {}
-    ENV.each_key {|k| @old_env[k] = ENV[k]}
-
-    if Gem::Version.new(`puppet --version`) >= Gem::Version.new('3.5')
-      Puppet.settings[:strict_variables]=true
-    end
-
-    Puppet.features.stubs(:root?).returns(true)
-  end
-
-  c.after :each do
-    PuppetlabsSpec::Files.cleanup
-  end
+  Puppet.settings[:strict_variables]=true Gem::Version.new(`puppet --version`) >= Gem::Version.new('3.5')
+  Puppet.features.stubs(:root?).returns(true)
 end
 
 require 'pathname'