Tests on puppet-openldap:

Before: Finished in 3 minutes 29.9 seconds (files took 0.57629 seconds to load)
After: Finished in 7.21 seconds (files took 1.15 seconds to load)

No comment... (exept this one :-) )